@@ -14,391 +14,276 @@ T _$identity<T>(T value) => value;
 
 /// @nodoc
 mixin _$Person {
-  String get id;
-  String get nickname;
-  String? get avatarPath;
-  bool get isDefault;
-  DateTime get createdAt;
 
-  /// Create a copy of Person
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $PersonCopyWith<Person> get copyWith =>
-      _$PersonCopyWithImpl<Person>(this as Person, _$identity);
+ String get id; String get nickname; String? get avatarPath; bool get isDefault; DateTime get createdAt;
+/// Create a copy of Person
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$PersonCopyWith<Person> get copyWith => _$PersonCopyWithImpl<Person>(this as Person, _$identity);
 
   /// Serializes this Person to a JSON map.
   Map<String, dynamic> toJson();
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is Person &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.nickname, nickname) ||
-                other.nickname == nickname) &&
-            (identical(other.avatarPath, avatarPath) ||
-                other.avatarPath == avatarPath) &&
-            (identical(other.isDefault, isDefault) ||
-                other.isDefault == isDefault) &&
-            (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt));
-  }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, id, nickname, avatarPath, isDefault, createdAt);
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is Person&&(identical(other.id, id) || other.id == id)&&(identical(other.nickname, nickname) || other.nickname == nickname)&&(identical(other.avatarPath, avatarPath) || other.avatarPath == avatarPath)&&(identical(other.isDefault, isDefault) || other.isDefault == isDefault)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+}
 
-  @override
-  String toString() {
-    return 'Person(id: $id, nickname: $nickname, avatarPath: $avatarPath, isDefault: $isDefault, createdAt: $createdAt)';
-  }
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,nickname,avatarPath,isDefault,createdAt);
+
+@override
+String toString() {
+  return 'Person(id: $id, nickname: $nickname, avatarPath: $avatarPath, isDefault: $isDefault, createdAt: $createdAt)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $PersonCopyWith<$Res> {
-  factory $PersonCopyWith(Person value, $Res Function(Person) _then) =
-      _$PersonCopyWithImpl;
-  @useResult
-  $Res call(
-      {String id,
-      String nickname,
-      String? avatarPath,
-      bool isDefault,
-      DateTime createdAt});
-}
+abstract mixin class $PersonCopyWith<$Res>  {
+  factory $PersonCopyWith(Person value, $Res Function(Person) _then) = _$PersonCopyWithImpl;
+@useResult
+$Res call({
+ String id, String nickname, String? avatarPath, bool isDefault, DateTime createdAt
+});
 
+
+
+
+}
 /// @nodoc
-class _$PersonCopyWithImpl<$Res> implements $PersonCopyWith<$Res> {
+class _$PersonCopyWithImpl<$Res>
+    implements $PersonCopyWith<$Res> {
   _$PersonCopyWithImpl(this._self, this._then);
 
   final Person _self;
   final $Res Function(Person) _then;
 
-  /// Create a copy of Person
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = null,
-    Object? nickname = null,
-    Object? avatarPath = freezed,
-    Object? isDefault = null,
-    Object? createdAt = null,
-  }) {
-    return _then(_self.copyWith(
-      id: null == id
-          ? _self.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      nickname: null == nickname
-          ? _self.nickname
-          : nickname // ignore: cast_nullable_to_non_nullable
-              as String,
-      avatarPath: freezed == avatarPath
-          ? _self.avatarPath
-          : avatarPath // ignore: cast_nullable_to_non_nullable
-              as String?,
-      isDefault: null == isDefault
-          ? _self.isDefault
-          : isDefault // ignore: cast_nullable_to_non_nullable
-              as bool,
-      createdAt: null == createdAt
-          ? _self.createdAt
-          : createdAt // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-    ));
-  }
+/// Create a copy of Person
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? nickname = null,Object? avatarPath = freezed,Object? isDefault = null,Object? createdAt = null,}) {
+  return _then(_self.copyWith(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,nickname: null == nickname ? _self.nickname : nickname // ignore: cast_nullable_to_non_nullable
+as String,avatarPath: freezed == avatarPath ? _self.avatarPath : avatarPath // ignore: cast_nullable_to_non_nullable
+as String?,isDefault: null == isDefault ? _self.isDefault : isDefault // ignore: cast_nullable_to_non_nullable
+as bool,createdAt: null == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
+as DateTime,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [Person].
 extension PersonPatterns on Person {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_Person value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Person() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _Person value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _Person() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_Person value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Person():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _Person value)  $default,){
+final _that = this;
+switch (_that) {
+case _Person():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_Person value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Person() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _Person value)?  $default,){
+final _that = this;
+switch (_that) {
+case _Person() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(String id, String nickname, String? avatarPath,
-            bool isDefault, DateTime createdAt)?
-        $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Person() when $default != null:
-        return $default(_that.id, _that.nickname, _that.avatarPath,
-            _that.isDefault, _that.createdAt);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String nickname,  String? avatarPath,  bool isDefault,  DateTime createdAt)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _Person() when $default != null:
+return $default(_that.id,_that.nickname,_that.avatarPath,_that.isDefault,_that.createdAt);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(String id, String nickname, String? avatarPath,
-            bool isDefault, DateTime createdAt)
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Person():
-        return $default(_that.id, _that.nickname, _that.avatarPath,
-            _that.isDefault, _that.createdAt);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String nickname,  String? avatarPath,  bool isDefault,  DateTime createdAt)  $default,) {final _that = this;
+switch (_that) {
+case _Person():
+return $default(_that.id,_that.nickname,_that.avatarPath,_that.isDefault,_that.createdAt);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(String id, String nickname, String? avatarPath,
-            bool isDefault, DateTime createdAt)?
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Person() when $default != null:
-        return $default(_that.id, _that.nickname, _that.avatarPath,
-            _that.isDefault, _that.createdAt);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String nickname,  String? avatarPath,  bool isDefault,  DateTime createdAt)?  $default,) {final _that = this;
+switch (_that) {
+case _Person() when $default != null:
+return $default(_that.id,_that.nickname,_that.avatarPath,_that.isDefault,_that.createdAt);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
 @JsonSerializable()
+
 class _Person implements Person {
-  const _Person(
-      {required this.id,
-      required this.nickname,
-      this.avatarPath,
-      this.isDefault = false,
-      required this.createdAt});
+  const _Person({required this.id, required this.nickname, this.avatarPath, this.isDefault = false, required this.createdAt});
   factory _Person.fromJson(Map<String, dynamic> json) => _$PersonFromJson(json);
 
-  @override
-  final String id;
-  @override
-  final String nickname;
-  @override
-  final String? avatarPath;
-  @override
-  @JsonKey()
-  final bool isDefault;
-  @override
-  final DateTime createdAt;
+@override final  String id;
+@override final  String nickname;
+@override final  String? avatarPath;
+@override@JsonKey() final  bool isDefault;
+@override final  DateTime createdAt;
 
-  /// Create a copy of Person
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$PersonCopyWith<_Person> get copyWith =>
-      __$PersonCopyWithImpl<_Person>(this, _$identity);
+/// Create a copy of Person
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$PersonCopyWith<_Person> get copyWith => __$PersonCopyWithImpl<_Person>(this, _$identity);
 
-  @override
-  Map<String, dynamic> toJson() {
-    return _$PersonToJson(
-      this,
-    );
-  }
+@override
+Map<String, dynamic> toJson() {
+  return _$PersonToJson(this, );
+}
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _Person &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.nickname, nickname) ||
-                other.nickname == nickname) &&
-            (identical(other.avatarPath, avatarPath) ||
-                other.avatarPath == avatarPath) &&
-            (identical(other.isDefault, isDefault) ||
-                other.isDefault == isDefault) &&
-            (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt));
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _Person&&(identical(other.id, id) || other.id == id)&&(identical(other.nickname, nickname) || other.nickname == nickname)&&(identical(other.avatarPath, avatarPath) || other.avatarPath == avatarPath)&&(identical(other.isDefault, isDefault) || other.isDefault == isDefault)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+}
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, id, nickname, avatarPath, isDefault, createdAt);
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,nickname,avatarPath,isDefault,createdAt);
 
-  @override
-  String toString() {
-    return 'Person(id: $id, nickname: $nickname, avatarPath: $avatarPath, isDefault: $isDefault, createdAt: $createdAt)';
-  }
+@override
+String toString() {
+  return 'Person(id: $id, nickname: $nickname, avatarPath: $avatarPath, isDefault: $isDefault, createdAt: $createdAt)';
+}
+
+
 }
 
 /// @nodoc
 abstract mixin class _$PersonCopyWith<$Res> implements $PersonCopyWith<$Res> {
-  factory _$PersonCopyWith(_Person value, $Res Function(_Person) _then) =
-      __$PersonCopyWithImpl;
-  @override
-  @useResult
-  $Res call(
-      {String id,
-      String nickname,
-      String? avatarPath,
-      bool isDefault,
-      DateTime createdAt});
-}
+  factory _$PersonCopyWith(_Person value, $Res Function(_Person) _then) = __$PersonCopyWithImpl;
+@override @useResult
+$Res call({
+ String id, String nickname, String? avatarPath, bool isDefault, DateTime createdAt
+});
 
+
+
+
+}
 /// @nodoc
-class __$PersonCopyWithImpl<$Res> implements _$PersonCopyWith<$Res> {
+class __$PersonCopyWithImpl<$Res>
+    implements _$PersonCopyWith<$Res> {
   __$PersonCopyWithImpl(this._self, this._then);
 
   final _Person _self;
   final $Res Function(_Person) _then;
 
-  /// Create a copy of Person
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? id = null,
-    Object? nickname = null,
-    Object? avatarPath = freezed,
-    Object? isDefault = null,
-    Object? createdAt = null,
-  }) {
-    return _then(_Person(
-      id: null == id
-          ? _self.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      nickname: null == nickname
-          ? _self.nickname
-          : nickname // ignore: cast_nullable_to_non_nullable
-              as String,
-      avatarPath: freezed == avatarPath
-          ? _self.avatarPath
-          : avatarPath // ignore: cast_nullable_to_non_nullable
-              as String?,
-      isDefault: null == isDefault
-          ? _self.isDefault
-          : isDefault // ignore: cast_nullable_to_non_nullable
-              as bool,
-      createdAt: null == createdAt
-          ? _self.createdAt
-          : createdAt // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-    ));
-  }
+/// Create a copy of Person
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? id = null,Object? nickname = null,Object? avatarPath = freezed,Object? isDefault = null,Object? createdAt = null,}) {
+  return _then(_Person(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,nickname: null == nickname ? _self.nickname : nickname // ignore: cast_nullable_to_non_nullable
+as String,avatarPath: freezed == avatarPath ? _self.avatarPath : avatarPath // ignore: cast_nullable_to_non_nullable
+as String?,isDefault: null == isDefault ? _self.isDefault : isDefault // ignore: cast_nullable_to_non_nullable
+as bool,createdAt: null == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
+as DateTime,
+  ));
+}
+
+
 }
 
 // dart format on
