@@ -14,406 +14,279 @@ T _$identity<T>(T value) => value;
 
 /// @nodoc
 mixin _$Tag {
-  String get id;
-  String? get personId;
-  String get name;
-  String get color;
-  bool get isSystem;
-  DateTime get createdAt;
 
-  /// Create a copy of Tag
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $TagCopyWith<Tag> get copyWith =>
-      _$TagCopyWithImpl<Tag>(this as Tag, _$identity);
+ String get id; String? get personId; String get name; String get color; bool get isSystem; DateTime get createdAt;
+/// Create a copy of Tag
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$TagCopyWith<Tag> get copyWith => _$TagCopyWithImpl<Tag>(this as Tag, _$identity);
 
   /// Serializes this Tag to a JSON map.
   Map<String, dynamic> toJson();
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is Tag &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.personId, personId) ||
-                other.personId == personId) &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.color, color) || other.color == color) &&
-            (identical(other.isSystem, isSystem) ||
-                other.isSystem == isSystem) &&
-            (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt));
-  }
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, id, personId, name, color, isSystem, createdAt);
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is Tag&&(identical(other.id, id) || other.id == id)&&(identical(other.personId, personId) || other.personId == personId)&&(identical(other.name, name) || other.name == name)&&(identical(other.color, color) || other.color == color)&&(identical(other.isSystem, isSystem) || other.isSystem == isSystem)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+}
 
-  @override
-  String toString() {
-    return 'Tag(id: $id, personId: $personId, name: $name, color: $color, isSystem: $isSystem, createdAt: $createdAt)';
-  }
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,personId,name,color,isSystem,createdAt);
+
+@override
+String toString() {
+  return 'Tag(id: $id, personId: $personId, name: $name, color: $color, isSystem: $isSystem, createdAt: $createdAt)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $TagCopyWith<$Res> {
+abstract mixin class $TagCopyWith<$Res>  {
   factory $TagCopyWith(Tag value, $Res Function(Tag) _then) = _$TagCopyWithImpl;
-  @useResult
-  $Res call(
-      {String id,
-      String? personId,
-      String name,
-      String color,
-      bool isSystem,
-      DateTime createdAt});
-}
+@useResult
+$Res call({
+ String id, String? personId, String name, String color, bool isSystem, DateTime createdAt
+});
 
+
+
+
+}
 /// @nodoc
-class _$TagCopyWithImpl<$Res> implements $TagCopyWith<$Res> {
+class _$TagCopyWithImpl<$Res>
+    implements $TagCopyWith<$Res> {
   _$TagCopyWithImpl(this._self, this._then);
 
   final Tag _self;
   final $Res Function(Tag) _then;
 
-  /// Create a copy of Tag
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = null,
-    Object? personId = freezed,
-    Object? name = null,
-    Object? color = null,
-    Object? isSystem = null,
-    Object? createdAt = null,
-  }) {
-    return _then(_self.copyWith(
-      id: null == id
-          ? _self.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      personId: freezed == personId
-          ? _self.personId
-          : personId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      name: null == name
-          ? _self.name
-          : name // ignore: cast_nullable_to_non_nullable
-              as String,
-      color: null == color
-          ? _self.color
-          : color // ignore: cast_nullable_to_non_nullable
-              as String,
-      isSystem: null == isSystem
-          ? _self.isSystem
-          : isSystem // ignore: cast_nullable_to_non_nullable
-              as bool,
-      createdAt: null == createdAt
-          ? _self.createdAt
-          : createdAt // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-    ));
-  }
+/// Create a copy of Tag
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? personId = freezed,Object? name = null,Object? color = null,Object? isSystem = null,Object? createdAt = null,}) {
+  return _then(_self.copyWith(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,personId: freezed == personId ? _self.personId : personId // ignore: cast_nullable_to_non_nullable
+as String?,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,color: null == color ? _self.color : color // ignore: cast_nullable_to_non_nullable
+as String,isSystem: null == isSystem ? _self.isSystem : isSystem // ignore: cast_nullable_to_non_nullable
+as bool,createdAt: null == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
+as DateTime,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [Tag].
 extension TagPatterns on Tag {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_Tag value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Tag() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _Tag value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _Tag() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_Tag value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Tag():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _Tag value)  $default,){
+final _that = this;
+switch (_that) {
+case _Tag():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_Tag value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Tag() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _Tag value)?  $default,){
+final _that = this;
+switch (_that) {
+case _Tag() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(String id, String? personId, String name, String color,
-            bool isSystem, DateTime createdAt)?
-        $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _Tag() when $default != null:
-        return $default(_that.id, _that.personId, _that.name, _that.color,
-            _that.isSystem, _that.createdAt);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String? personId,  String name,  String color,  bool isSystem,  DateTime createdAt)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _Tag() when $default != null:
+return $default(_that.id,_that.personId,_that.name,_that.color,_that.isSystem,_that.createdAt);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(String id, String? personId, String name, String color,
-            bool isSystem, DateTime createdAt)
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Tag():
-        return $default(_that.id, _that.personId, _that.name, _that.color,
-            _that.isSystem, _that.createdAt);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String? personId,  String name,  String color,  bool isSystem,  DateTime createdAt)  $default,) {final _that = this;
+switch (_that) {
+case _Tag():
+return $default(_that.id,_that.personId,_that.name,_that.color,_that.isSystem,_that.createdAt);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(String id, String? personId, String name, String color,
-            bool isSystem, DateTime createdAt)?
-        $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _Tag() when $default != null:
-        return $default(_that.id, _that.personId, _that.name, _that.color,
-            _that.isSystem, _that.createdAt);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String? personId,  String name,  String color,  bool isSystem,  DateTime createdAt)?  $default,) {final _that = this;
+switch (_that) {
+case _Tag() when $default != null:
+return $default(_that.id,_that.personId,_that.name,_that.color,_that.isSystem,_that.createdAt);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
 @JsonSerializable()
+
 class _Tag implements Tag {
-  const _Tag(
-      {required this.id,
-      this.personId,
-      required this.name,
-      required this.color,
-      this.isSystem = false,
-      required this.createdAt});
+  const _Tag({required this.id, this.personId, required this.name, required this.color, this.isSystem = false, required this.createdAt});
   factory _Tag.fromJson(Map<String, dynamic> json) => _$TagFromJson(json);
 
-  @override
-  final String id;
-  @override
-  final String? personId;
-  @override
-  final String name;
-  @override
-  final String color;
-  @override
-  @JsonKey()
-  final bool isSystem;
-  @override
-  final DateTime createdAt;
+@override final  String id;
+@override final  String? personId;
+@override final  String name;
+@override final  String color;
+@override@JsonKey() final  bool isSystem;
+@override final  DateTime createdAt;
 
-  /// Create a copy of Tag
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$TagCopyWith<_Tag> get copyWith =>
-      __$TagCopyWithImpl<_Tag>(this, _$identity);
+/// Create a copy of Tag
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$TagCopyWith<_Tag> get copyWith => __$TagCopyWithImpl<_Tag>(this, _$identity);
 
-  @override
-  Map<String, dynamic> toJson() {
-    return _$TagToJson(
-      this,
-    );
-  }
+@override
+Map<String, dynamic> toJson() {
+  return _$TagToJson(this, );
+}
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _Tag &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.personId, personId) ||
-                other.personId == personId) &&
-            (identical(other.name, name) || other.name == name) &&
-            (identical(other.color, color) || other.color == color) &&
-            (identical(other.isSystem, isSystem) ||
-                other.isSystem == isSystem) &&
-            (identical(other.createdAt, createdAt) ||
-                other.createdAt == createdAt));
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _Tag&&(identical(other.id, id) || other.id == id)&&(identical(other.personId, personId) || other.personId == personId)&&(identical(other.name, name) || other.name == name)&&(identical(other.color, color) || other.color == color)&&(identical(other.isSystem, isSystem) || other.isSystem == isSystem)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+}
 
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  int get hashCode =>
-      Object.hash(runtimeType, id, personId, name, color, isSystem, createdAt);
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,personId,name,color,isSystem,createdAt);
 
-  @override
-  String toString() {
-    return 'Tag(id: $id, personId: $personId, name: $name, color: $color, isSystem: $isSystem, createdAt: $createdAt)';
-  }
+@override
+String toString() {
+  return 'Tag(id: $id, personId: $personId, name: $name, color: $color, isSystem: $isSystem, createdAt: $createdAt)';
+}
+
+
 }
 
 /// @nodoc
 abstract mixin class _$TagCopyWith<$Res> implements $TagCopyWith<$Res> {
-  factory _$TagCopyWith(_Tag value, $Res Function(_Tag) _then) =
-      __$TagCopyWithImpl;
-  @override
-  @useResult
-  $Res call(
-      {String id,
-      String? personId,
-      String name,
-      String color,
-      bool isSystem,
-      DateTime createdAt});
-}
+  factory _$TagCopyWith(_Tag value, $Res Function(_Tag) _then) = __$TagCopyWithImpl;
+@override @useResult
+$Res call({
+ String id, String? personId, String name, String color, bool isSystem, DateTime createdAt
+});
 
+
+
+
+}
 /// @nodoc
-class __$TagCopyWithImpl<$Res> implements _$TagCopyWith<$Res> {
+class __$TagCopyWithImpl<$Res>
+    implements _$TagCopyWith<$Res> {
   __$TagCopyWithImpl(this._self, this._then);
 
   final _Tag _self;
   final $Res Function(_Tag) _then;
 
-  /// Create a copy of Tag
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? id = null,
-    Object? personId = freezed,
-    Object? name = null,
-    Object? color = null,
-    Object? isSystem = null,
-    Object? createdAt = null,
-  }) {
-    return _then(_Tag(
-      id: null == id
-          ? _self.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      personId: freezed == personId
-          ? _self.personId
-          : personId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      name: null == name
-          ? _self.name
-          : name // ignore: cast_nullable_to_non_nullable
-              as String,
-      color: null == color
-          ? _self.color
-          : color // ignore: cast_nullable_to_non_nullable
-              as String,
-      isSystem: null == isSystem
-          ? _self.isSystem
-          : isSystem // ignore: cast_nullable_to_non_nullable
-              as bool,
-      createdAt: null == createdAt
-          ? _self.createdAt
-          : createdAt // ignore: cast_nullable_to_non_nullable
-              as DateTime,
-    ));
-  }
+/// Create a copy of Tag
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? id = null,Object? personId = freezed,Object? name = null,Object? color = null,Object? isSystem = null,Object? createdAt = null,}) {
+  return _then(_Tag(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,personId: freezed == personId ? _self.personId : personId // ignore: cast_nullable_to_non_nullable
+as String?,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,color: null == color ? _self.color : color // ignore: cast_nullable_to_non_nullable
+as String,isSystem: null == isSystem ? _self.isSystem : isSystem // ignore: cast_nullable_to_non_nullable
+as bool,createdAt: null == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
+as DateTime,
+  ));
+}
+
+
 }
 
 // dart format on
