@@ -7,29 +7,30 @@ part of 'ocr_result.dart';
 // **************************************************************************
 
 _OCRBlock _$OCRBlockFromJson(Map<String, dynamic> json) => _OCRBlock(
-      text: json['text'] as String,
-      left: (json['left'] as num).toDouble(),
-      top: (json['top'] as num).toDouble(),
-      width: (json['width'] as num).toDouble(),
-      height: (json['height'] as num).toDouble(),
-    );
+  text: json['text'] as String,
+  left: (json['left'] as num).toDouble(),
+  top: (json['top'] as num).toDouble(),
+  width: (json['width'] as num).toDouble(),
+  height: (json['height'] as num).toDouble(),
+);
 
 Map<String, dynamic> _$OCRBlockToJson(_OCRBlock instance) => <String, dynamic>{
-      'text': instance.text,
-      'left': instance.left,
-      'top': instance.top,
-      'width': instance.width,
-      'height': instance.height,
-    };
+  'text': instance.text,
+  'left': instance.left,
+  'top': instance.top,
+  'width': instance.width,
+  'height': instance.height,
+};
 
 _OCRResult _$OCRResultFromJson(Map<String, dynamic> json) => _OCRResult(
-      text: json['text'] as String,
-      blocks: (json['blocks'] as List<dynamic>?)
-              ?.map((e) => OCRBlock.fromJson(e as Map<String, dynamic>))
-              .toList() ??
-          const [],
-      confidence: (json['confidence'] as num?)?.toDouble() ?? 0.0,
-    );
+  text: json['text'] as String,
+  blocks:
+      (json['blocks'] as List<dynamic>?)
+          ?.map((e) => OCRBlock.fromJson(e as Map<String, dynamic>))
+          .toList() ??
+      const [],
+  confidence: (json['confidence'] as num?)?.toDouble() ?? 0.0,
+);
 
 Map<String, dynamic> _$OCRResultToJson(_OCRResult instance) =>
     <String, dynamic>{
