@@ -11,298 +11,256 @@ part of 'search_result.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
 /// @nodoc
 mixin _$SearchResult {
-  MedicalRecord get record;
 
-  /// 包含高亮 HTML 标签的摘要
-  String get snippet;
+ MedicalRecord get record;/// 包含高亮 HTML 标签的摘要
+ String get snippet;
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$SearchResultCopyWith<SearchResult> get copyWith => _$SearchResultCopyWithImpl<SearchResult>(this as SearchResult, _$identity);
 
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $SearchResultCopyWith<SearchResult> get copyWith =>
-      _$SearchResultCopyWithImpl<SearchResult>(
-          this as SearchResult, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is SearchResult &&
-            (identical(other.record, record) || other.record == record) &&
-            (identical(other.snippet, snippet) || other.snippet == snippet));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, record, snippet);
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is SearchResult&&(identical(other.record, record) || other.record == record)&&(identical(other.snippet, snippet) || other.snippet == snippet));
+}
 
-  @override
-  String toString() {
-    return 'SearchResult(record: $record, snippet: $snippet)';
-  }
+
+@override
+int get hashCode => Object.hash(runtimeType,record,snippet);
+
+@override
+String toString() {
+  return 'SearchResult(record: $record, snippet: $snippet)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $SearchResultCopyWith<$Res> {
-  factory $SearchResultCopyWith(
-          SearchResult value, $Res Function(SearchResult) _then) =
-      _$SearchResultCopyWithImpl;
-  @useResult
-  $Res call({MedicalRecord record, String snippet});
+abstract mixin class $SearchResultCopyWith<$Res>  {
+  factory $SearchResultCopyWith(SearchResult value, $Res Function(SearchResult) _then) = _$SearchResultCopyWithImpl;
+@useResult
+$Res call({
+ MedicalRecord record, String snippet
+});
 
-  $MedicalRecordCopyWith<$Res> get record;
+
+$MedicalRecordCopyWith<$Res> get record;
+
 }
-
 /// @nodoc
-class _$SearchResultCopyWithImpl<$Res> implements $SearchResultCopyWith<$Res> {
+class _$SearchResultCopyWithImpl<$Res>
+    implements $SearchResultCopyWith<$Res> {
   _$SearchResultCopyWithImpl(this._self, this._then);
 
   final SearchResult _self;
   final $Res Function(SearchResult) _then;
 
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? record = null,
-    Object? snippet = null,
-  }) {
-    return _then(_self.copyWith(
-      record: null == record
-          ? _self.record
-          : record // ignore: cast_nullable_to_non_nullable
-              as MedicalRecord,
-      snippet: null == snippet
-          ? _self.snippet
-          : snippet // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $MedicalRecordCopyWith<$Res> get record {
-    return $MedicalRecordCopyWith<$Res>(_self.record, (value) {
-      return _then(_self.copyWith(record: value));
-    });
-  }
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? record = null,Object? snippet = null,}) {
+  return _then(_self.copyWith(
+record: null == record ? _self.record : record // ignore: cast_nullable_to_non_nullable
+as MedicalRecord,snippet: null == snippet ? _self.snippet : snippet // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
 }
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$MedicalRecordCopyWith<$Res> get record {
+  
+  return $MedicalRecordCopyWith<$Res>(_self.record, (value) {
+    return _then(_self.copyWith(record: value));
+  });
+}
+}
+
 
 /// Adds pattern-matching-related methods to [SearchResult].
 extension SearchResultPatterns on SearchResult {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_SearchResult value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _SearchResult value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _SearchResult() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_SearchResult value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _SearchResult value)  $default,){
+final _that = this;
+switch (_that) {
+case _SearchResult():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_SearchResult value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _SearchResult value)?  $default,){
+final _that = this;
+switch (_that) {
+case _SearchResult() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(MedicalRecord record, String snippet)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult() when $default != null:
-        return $default(_that.record, _that.snippet);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( MedicalRecord record,  String snippet)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _SearchResult() when $default != null:
+return $default(_that.record,_that.snippet);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(MedicalRecord record, String snippet) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult():
-        return $default(_that.record, _that.snippet);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( MedicalRecord record,  String snippet)  $default,) {final _that = this;
+switch (_that) {
+case _SearchResult():
+return $default(_that.record,_that.snippet);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(MedicalRecord record, String snippet)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _SearchResult() when $default != null:
-        return $default(_that.record, _that.snippet);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( MedicalRecord record,  String snippet)?  $default,) {final _that = this;
+switch (_that) {
+case _SearchResult() when $default != null:
+return $default(_that.record,_that.snippet);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _SearchResult implements SearchResult {
   const _SearchResult({required this.record, required this.snippet});
+  
 
-  @override
-  final MedicalRecord record;
+@override final  MedicalRecord record;
+/// 包含高亮 HTML 标签的摘要
+@override final  String snippet;
 
-  /// 包含高亮 HTML 标签的摘要
-  @override
-  final String snippet;
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$SearchResultCopyWith<_SearchResult> get copyWith => __$SearchResultCopyWithImpl<_SearchResult>(this, _$identity);
 
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$SearchResultCopyWith<_SearchResult> get copyWith =>
-      __$SearchResultCopyWithImpl<_SearchResult>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _SearchResult &&
-            (identical(other.record, record) || other.record == record) &&
-            (identical(other.snippet, snippet) || other.snippet == snippet));
-  }
 
-  @override
-  int get hashCode => Object.hash(runtimeType, record, snippet);
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _SearchResult&&(identical(other.record, record) || other.record == record)&&(identical(other.snippet, snippet) || other.snippet == snippet));
+}
 
-  @override
-  String toString() {
-    return 'SearchResult(record: $record, snippet: $snippet)';
-  }
+
+@override
+int get hashCode => Object.hash(runtimeType,record,snippet);
+
+@override
+String toString() {
+  return 'SearchResult(record: $record, snippet: $snippet)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class _$SearchResultCopyWith<$Res>
-    implements $SearchResultCopyWith<$Res> {
-  factory _$SearchResultCopyWith(
-          _SearchResult value, $Res Function(_SearchResult) _then) =
-      __$SearchResultCopyWithImpl;
-  @override
-  @useResult
-  $Res call({MedicalRecord record, String snippet});
+abstract mixin class _$SearchResultCopyWith<$Res> implements $SearchResultCopyWith<$Res> {
+  factory _$SearchResultCopyWith(_SearchResult value, $Res Function(_SearchResult) _then) = __$SearchResultCopyWithImpl;
+@override @useResult
+$Res call({
+ MedicalRecord record, String snippet
+});
 
-  @override
-  $MedicalRecordCopyWith<$Res> get record;
+
+@override $MedicalRecordCopyWith<$Res> get record;
+
 }
-
 /// @nodoc
 class __$SearchResultCopyWithImpl<$Res>
     implements _$SearchResultCopyWith<$Res> {
@@ -311,35 +269,26 @@ class __$SearchResultCopyWithImpl<$Res>
   final _SearchResult _self;
   final $Res Function(_SearchResult) _then;
 
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? record = null,
-    Object? snippet = null,
-  }) {
-    return _then(_SearchResult(
-      record: null == record
-          ? _self.record
-          : record // ignore: cast_nullable_to_non_nullable
-              as MedicalRecord,
-      snippet: null == snippet
-          ? _self.snippet
-          : snippet // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? record = null,Object? snippet = null,}) {
+  return _then(_SearchResult(
+record: null == record ? _self.record : record // ignore: cast_nullable_to_non_nullable
+as MedicalRecord,snippet: null == snippet ? _self.snippet : snippet // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
+}
 
-  /// Create a copy of SearchResult
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $MedicalRecordCopyWith<$Res> get record {
-    return $MedicalRecordCopyWith<$Res>(_self.record, (value) {
-      return _then(_self.copyWith(record: value));
-    });
-  }
+/// Create a copy of SearchResult
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$MedicalRecordCopyWith<$Res> get record {
+  
+  return $MedicalRecordCopyWith<$Res>(_self.record, (value) {
+    return _then(_self.copyWith(record: value));
+  });
+}
 }
 
 // dart format on
