@@ -15,15 +15,15 @@ final reviewListControllerProvider = ReviewListControllerProvider._();
 final class ReviewListControllerProvider
     extends $AsyncNotifierProvider<ReviewListController, List<MedicalRecord>> {
   ReviewListControllerProvider._()
-      : super(
-          from: null,
-          argument: null,
-          retry: null,
-          name: r'reviewListControllerProvider',
-          isAutoDispose: true,
-          dependencies: null,
-          $allTransitiveDependencies: null,
-        );
+    : super(
+        from: null,
+        argument: null,
+        retry: null,
+        name: r'reviewListControllerProvider',
+        isAutoDispose: true,
+        dependencies: null,
+        $allTransitiveDependencies: null,
+      );
 
   @override
   String debugGetCreateSourceHash() => _$reviewListControllerHash();
@@ -44,11 +44,14 @@ abstract class _$ReviewListController
   void runBuild() {
     final ref =
         this.ref as $Ref<AsyncValue<List<MedicalRecord>>, List<MedicalRecord>>;
-    final element = ref.element as $ClassProviderElement<
-        AnyNotifier<AsyncValue<List<MedicalRecord>>, List<MedicalRecord>>,
-        AsyncValue<List<MedicalRecord>>,
-        Object?,
-        Object?>;
+    final element =
+        ref.element
+            as $ClassProviderElement<
+              AnyNotifier<AsyncValue<List<MedicalRecord>>, List<MedicalRecord>>,
+              AsyncValue<List<MedicalRecord>>,
+              Object?,
+              Object?
+            >;
     element.handleCreate(ref, build);
   }
 }
