@@ -11,317 +11,270 @@ part of 'home_state.dart';
 
 // dart format off
 T _$identity<T>(T value) => value;
-
 /// @nodoc
 mixin _$HomeState {
-  List<MedicalRecord> get records;
-  int get pendingCount;
 
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  $HomeStateCopyWith<HomeState> get copyWith =>
-      _$HomeStateCopyWithImpl<HomeState>(this as HomeState, _$identity);
+ List<MedicalRecord> get records; int get pendingCount;
+/// Create a copy of HomeState
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$HomeStateCopyWith<HomeState> get copyWith => _$HomeStateCopyWithImpl<HomeState>(this as HomeState, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is HomeState &&
-            const DeepCollectionEquality().equals(other.records, records) &&
-            (identical(other.pendingCount, pendingCount) ||
-                other.pendingCount == pendingCount));
-  }
 
-  @override
-  int get hashCode => Object.hash(
-      runtimeType, const DeepCollectionEquality().hash(records), pendingCount);
 
-  @override
-  String toString() {
-    return 'HomeState(records: $records, pendingCount: $pendingCount)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is HomeState&&const DeepCollectionEquality().equals(other.records, records)&&(identical(other.pendingCount, pendingCount) || other.pendingCount == pendingCount));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(records),pendingCount);
+
+@override
+String toString() {
+  return 'HomeState(records: $records, pendingCount: $pendingCount)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class $HomeStateCopyWith<$Res> {
-  factory $HomeStateCopyWith(HomeState value, $Res Function(HomeState) _then) =
-      _$HomeStateCopyWithImpl;
-  @useResult
-  $Res call({List<MedicalRecord> records, int pendingCount});
-}
+abstract mixin class $HomeStateCopyWith<$Res>  {
+  factory $HomeStateCopyWith(HomeState value, $Res Function(HomeState) _then) = _$HomeStateCopyWithImpl;
+@useResult
+$Res call({
+ List<MedicalRecord> records, int pendingCount
+});
 
+
+
+
+}
 /// @nodoc
-class _$HomeStateCopyWithImpl<$Res> implements $HomeStateCopyWith<$Res> {
+class _$HomeStateCopyWithImpl<$Res>
+    implements $HomeStateCopyWith<$Res> {
   _$HomeStateCopyWithImpl(this._self, this._then);
 
   final HomeState _self;
   final $Res Function(HomeState) _then;
 
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? records = null,
-    Object? pendingCount = null,
-  }) {
-    return _then(_self.copyWith(
-      records: null == records
-          ? _self.records
-          : records // ignore: cast_nullable_to_non_nullable
-              as List<MedicalRecord>,
-      pendingCount: null == pendingCount
-          ? _self.pendingCount
-          : pendingCount // ignore: cast_nullable_to_non_nullable
-              as int,
-    ));
-  }
+/// Create a copy of HomeState
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? records = null,Object? pendingCount = null,}) {
+  return _then(_self.copyWith(
+records: null == records ? _self.records : records // ignore: cast_nullable_to_non_nullable
+as List<MedicalRecord>,pendingCount: null == pendingCount ? _self.pendingCount : pendingCount // ignore: cast_nullable_to_non_nullable
+as int,
+  ));
 }
+
+}
+
 
 /// Adds pattern-matching-related methods to [HomeState].
 extension HomeStatePatterns on HomeState {
-  /// A variant of `map` that fallback to returning `orElse`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>(
-    TResult Function(_HomeState value)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState() when $default != null:
-        return $default(_that);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _HomeState value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _HomeState() when $default != null:
+return $default(_that);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// Callbacks receives the raw object, upcasted.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case final Subclass2 value:
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>(
-    TResult Function(_HomeState value) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState():
-        return $default(_that);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _HomeState value)  $default,){
+final _that = this;
+switch (_that) {
+case _HomeState():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `map` that fallback to returning `null`.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case final Subclass value:
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>(
-    TResult? Function(_HomeState value)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState() when $default != null:
-        return $default(_that);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _HomeState value)?  $default,){
+final _that = this;
+switch (_that) {
+case _HomeState() when $default != null:
+return $default(_that);case _:
+  return null;
 
-  /// A variant of `when` that fallback to an `orElse` callback.
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return orElse();
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>(
-    TResult Function(List<MedicalRecord> records, int pendingCount)? $default, {
-    required TResult orElse(),
-  }) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState() when $default != null:
-        return $default(_that.records, _that.pendingCount);
-      case _:
-        return orElse();
-    }
-  }
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( List<MedicalRecord> records,  int pendingCount)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _HomeState() when $default != null:
+return $default(_that.records,_that.pendingCount);case _:
+  return orElse();
 
-  /// A `switch`-like method, using callbacks.
-  ///
-  /// As opposed to `map`, this offers destructuring.
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case Subclass2(:final field2):
-  ///     return ...;
-  /// }
-  /// ```
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>(
-    TResult Function(List<MedicalRecord> records, int pendingCount) $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState():
-        return $default(_that.records, _that.pendingCount);
-      case _:
-        throw StateError('Unexpected subclass');
-    }
-  }
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( List<MedicalRecord> records,  int pendingCount)  $default,) {final _that = this;
+switch (_that) {
+case _HomeState():
+return $default(_that.records,_that.pendingCount);case _:
+  throw StateError('Unexpected subclass');
 
-  /// A variant of `when` that fallback to returning `null`
-  ///
-  /// It is equivalent to doing:
-  /// ```dart
-  /// switch (sealedClass) {
-  ///   case Subclass(:final field):
-  ///     return ...;
-  ///   case _:
-  ///     return null;
-  /// }
-  /// ```
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
 
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>(
-    TResult? Function(List<MedicalRecord> records, int pendingCount)? $default,
-  ) {
-    final _that = this;
-    switch (_that) {
-      case _HomeState() when $default != null:
-        return $default(_that.records, _that.pendingCount);
-      case _:
-        return null;
-    }
-  }
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( List<MedicalRecord> records,  int pendingCount)?  $default,) {final _that = this;
+switch (_that) {
+case _HomeState() when $default != null:
+return $default(_that.records,_that.pendingCount);case _:
+  return null;
+
+}
+}
+
 }
 
 /// @nodoc
+
 
 class _HomeState implements HomeState {
-  const _HomeState(
-      {final List<MedicalRecord> records = const [], this.pendingCount = 0})
-      : _records = records;
+  const _HomeState({final  List<MedicalRecord> records = const [], this.pendingCount = 0}): _records = records;
+  
 
-  final List<MedicalRecord> _records;
-  @override
-  @JsonKey()
-  List<MedicalRecord> get records {
-    if (_records is EqualUnmodifiableListView) return _records;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(_records);
-  }
+ final  List<MedicalRecord> _records;
+@override@JsonKey() List<MedicalRecord> get records {
+  if (_records is EqualUnmodifiableListView) return _records;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(_records);
+}
 
-  @override
-  @JsonKey()
-  final int pendingCount;
+@override@JsonKey() final  int pendingCount;
 
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @pragma('vm:prefer-inline')
-  _$HomeStateCopyWith<_HomeState> get copyWith =>
-      __$HomeStateCopyWithImpl<_HomeState>(this, _$identity);
+/// Create a copy of HomeState
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$HomeStateCopyWith<_HomeState> get copyWith => __$HomeStateCopyWithImpl<_HomeState>(this, _$identity);
 
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _HomeState &&
-            const DeepCollectionEquality().equals(other._records, _records) &&
-            (identical(other.pendingCount, pendingCount) ||
-                other.pendingCount == pendingCount));
-  }
 
-  @override
-  int get hashCode => Object.hash(
-      runtimeType, const DeepCollectionEquality().hash(_records), pendingCount);
 
-  @override
-  String toString() {
-    return 'HomeState(records: $records, pendingCount: $pendingCount)';
-  }
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _HomeState&&const DeepCollectionEquality().equals(other._records, _records)&&(identical(other.pendingCount, pendingCount) || other.pendingCount == pendingCount));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(_records),pendingCount);
+
+@override
+String toString() {
+  return 'HomeState(records: $records, pendingCount: $pendingCount)';
+}
+
+
 }
 
 /// @nodoc
-abstract mixin class _$HomeStateCopyWith<$Res>
-    implements $HomeStateCopyWith<$Res> {
-  factory _$HomeStateCopyWith(
-          _HomeState value, $Res Function(_HomeState) _then) =
-      __$HomeStateCopyWithImpl;
-  @override
-  @useResult
-  $Res call({List<MedicalRecord> records, int pendingCount});
-}
+abstract mixin class _$HomeStateCopyWith<$Res> implements $HomeStateCopyWith<$Res> {
+  factory _$HomeStateCopyWith(_HomeState value, $Res Function(_HomeState) _then) = __$HomeStateCopyWithImpl;
+@override @useResult
+$Res call({
+ List<MedicalRecord> records, int pendingCount
+});
 
+
+
+
+}
 /// @nodoc
-class __$HomeStateCopyWithImpl<$Res> implements _$HomeStateCopyWith<$Res> {
+class __$HomeStateCopyWithImpl<$Res>
+    implements _$HomeStateCopyWith<$Res> {
   __$HomeStateCopyWithImpl(this._self, this._then);
 
   final _HomeState _self;
   final $Res Function(_HomeState) _then;
 
-  /// Create a copy of HomeState
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @pragma('vm:prefer-inline')
-  $Res call({
-    Object? records = null,
-    Object? pendingCount = null,
-  }) {
-    return _then(_HomeState(
-      records: null == records
-          ? _self._records
-          : records // ignore: cast_nullable_to_non_nullable
-              as List<MedicalRecord>,
-      pendingCount: null == pendingCount
-          ? _self.pendingCount
-          : pendingCount // ignore: cast_nullable_to_non_nullable
-              as int,
-    ));
-  }
+/// Create a copy of HomeState
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? records = null,Object? pendingCount = null,}) {
+  return _then(_HomeState(
+records: null == records ? _self._records : records // ignore: cast_nullable_to_non_nullable
+as List<MedicalRecord>,pendingCount: null == pendingCount ? _self.pendingCount : pendingCount // ignore: cast_nullable_to_non_nullable
+as int,
+  ));
+}
+
+
 }
 
 // dart format on
